[Syntax] Fix ending curly brace not getting highlighted.
This commit fixes ending curly brace not getting highlighted. The
`expression` context of JavaScript actually sets into a series of context
hence the `edge_expression_body` was popped off the stack earlier than
expected. The solution is to just push another context and then include
the `expression` context so that when the `expression` context sets, it
ends up back in the `edge_expression_body`.